Explicitly link all necessary deps for VST plugin on Linux
Otherwise, the linker silently allows missing symbol references.
That makes it very unreliable across VST hosts because the ones
that have already loaded the missing symbols will work but those
that have not fail to load.

This makes the plugin work also in Carla. Still not working in
Audacity, but then again no VST3 plugins seem to work with it,
even the VST3 SDK examples!

Add support for setBusArrangement() to enforce either mono -> mono
or stereo -> stereo. Otherwise, the plugin appears to support 64
inputs and 64 outputs, which is incorrect.