Fix bug preventing logout from virtual studio mode
It actually was logging out by clearing the access token, but we
were failing to also clear the new refresh token. So it would just
immediately trigger the cycle to use the refresh token to get a
new access token, effectively logging in again right away.

Clearing the refresh token on logout fixes the problem, but also
best to stop the refresh timer (it will get started again after
successful login).

Also bumping version and changelog for beta2 release