Replacing Receive and Passthrough knobs with Mix and Gain
I've learned that trying to use two sliders to control the mixing
of two channels of audio is always a bad idea. It's always better
to use a single knob or slider that adjusts the mix between the
two, and if necessary a separate gain knob.

Also fixing double trigger of vs token refresh at startup.