Removing the extra Gain dial from audio bridge interface
Jan and I went through a bunch of iterations on this, and came
to the realization that there is really no need to applying
extra gain to the plugin's output audio, and the extra UX
complexity this third dial adds wasn't worth it.

This updated design also adds some extra labels to help make
it clearer where the "Send" is going, that the "Mix" applies
to audio output from the plugin, and what you are mixing between.