Temporarily disabling support for feedback detection in meson builds
I keep coming across crashes with backtraces pointing to
`Analyzer::onTick`. Also, this feature never really worked all that
well: I regularly hear complaints about it triggering when it shouldn't
and I've also seen it not trigger when it should. I think it's best to
just remove this until we have time to put more effort into it.