Adding Audio Unit v2 version of the JackTrip Audio Bridge
Updated packaging to include Audio Unit for OSX, and share
resources with the VST3 plugin. Both plugins now share a
common AudioBridgeProcessor class that implements more of
the plugin-oriented functionality.

A v3 audio unit version is also included in the src/auv3
directory, but it is not currently working. It's
extremely difficult to get MacOS to recognize these app
extensions, and there seems to be no documentation or
debugging tools available. Getting it to work is going to
require a lot more trial and error hacking than I have
time for right now, and there appears to be few -- if
any -- real benefits to using v3. And at least according
to the AIs, many hosts don't even support v3 yet.