When skipping packets, advance last and use next for training
Previously, we would use the next good packet for training, but
never use it as "real" packet. Now if at least one packet is being
skipped, we will use the next good packet for training but only
advance the last pointer to the slot before it. This helps to
ensure that the next pull will return the next good packet.

My testing so far indicates that this significantly improves
clarity when there is packet loss or delay spikes.