Updating AudioSockets to eliminate dependency on QApplication
Creating a static QApplication and using QThread's EventLoop with
signals and slots doesn't seem to work well in the context of
plugins that can be loaded and unloaded dynamically.

In particular, I encountered some problems with Windows VST hosts
where the first plugin you create works OK. But if you delete it
and then create a second one, it locks up. It appears that there
is some static flag getting set deep in QThread's EventLoop
processing that basically just makes all threads exit.