Don't update mXferBuffer when processing packet as a glitch
bugfix/plc-skipped-training started updating this whenever we had
a valid packet but processed a glitch due to tolerance being
exceeded. Burg seems to use this buffer even when the glitch flag
is true, so I had thought that maybe we should keep it updated
with the latest known value?

This reverts behavior to only update mXferBuffer when a packet
is not being processed as a glitch. It will stay the same value
for any sequential glitch processing.