Only use glitches for increasing auto headroom
Requiring a skipped packets that exceeded tolerance by less
than past millisecond was kind of arbitrary, and we want to
start preferring quality over latency whenever auto is
enabled.

There is an edge case where if someone has sustained packet
loss (say 1%) due to a really bad connection, it's going to
keep bumping headroom until it hits max allowed. But that
seems quite rare and if it is the case, they can always
use the slider to adjust.

Updated max headroom calculation to use 10ms greater than
the last good packet used, as opposed to a fixed value of
100ms. That was also quite arbitrary. This is perhaps
slightly less so?

Ensure that we don't log in Regulator unless gVerbose is
set (-V). Some of these happen inside the audio callback
thread, which is one place you should never log outside
of testing.

Updated the ChangeDevices page to clean up the Auto button
for the Audio quality slider a little bit. It was kind of
jumpy and awkward previously.

Added a RtAudioInterface::sampleRateChanged method to try
and detect this, but the condition doesn't seem to be
getting detected by RtAudio..